requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
Jinja2==3.1.2
mysql-connector-python==8.0.33
Flask-MySQL==1.5.2 